Move Status to the own namespace
